+++ freemarker-generator-tools/src/test/java/org/apache/freemarker/generator/tools/jsoup/JsoupToolTest.java
@@ -54,6 +54,4 @@ public class JsoupToolTest {
     private JsoupTool jsoupTool() {
         return new JsoupTool();
     }
-
-
 }